requirements.txt
aiogram>=3.22,<4
aiosqlite>=0.20
python-dotenv>=1.0
gspread>=6.0
google-auth>=2.0
